feat: generalise notion of program in bytecode interpreter (#1924)
feat: refactor bytecode IR

This refactors the bytecode IR so that it matches WIR almost
identically.  The step is a precursor before replacing WIR altogether.
Amongst other things, this makes a strong separation between a bytecode
"assembly" program and a compiled bytecode program.  The latter only
exists within the realm of the interpreter, whilst the former presents a
near-identical API to WIR.  A notion of "descriptor" is introduced to
replace the existing WIR notions of function and memory, and also
register.Register.  This decouples the bytecode IR from the legacy (i.e.
MIR/AIR) layers.  It also provides a clear separation between what is a
program-level construct (i.e. a descriptor) and what is an
execution-level construct (i.e. an interpreter).

Descriptors.  A new pkg/zkc/vm/internal/descriptor package captures
everything needed to describe a program without reference to the legacy
layers: a Program of Modules, where each Module is either a Function or a
Memory, plus a register abstraction which replaces the use of
register.Register.  The word-to-bytecode compiler now lowers a word
machine into a descriptor.Program of these constructs, rather than
reaching into WIR types directly.

Encoding.  All encoding/decoding logic is lifted out of the bytecode
package and into a new pkg/zkc/vm/internal/interpreter/encoding package,
sitting alongside the interpreter itself.  After this move the bytecode
package holds only the "assembly" representation: each instruction is a
self-describing value (arith, bitwise, cat, call, checkcast, debug, div,
fail, field, jmp, ret, skip, skip_if, switch) wrapped by a vector, which
is what gives it its near-identical shape to WIR.  The obsolete machinery
that conflated the two roles -- the old encoder, in-package program and
memory-map types -- is removed.

Incidental changes:

  - A generic, overflow-checked util.Cast helper (pkg/util/cast.go) is
    added for narrowing numeric conversions and used throughout the new
    code in place of ad-hoc casts.

  - The word-to-bytecode transform flattens the per-instruction bytecode
    "packets" emitted during compilation into a single stream,
    recalculating branch targets (Skip, SkipIf and Switch) so they remain
    expressed in bytecodes rather than packets.

  - A new switch_10 unit test exercises multiway dispatch end to end.

+++ pkg/zkc/vm/internal/checkpoint/memory.go
@@ -6,20 +6,20 @@ package checkpoint
 // are implicitly zero.
 type Memory[W any] struct {
 	// Module identifier for this memory.
-	moduleId uint
+	moduleId uint16
 	// Pages determines the contents of the given memory in this snapshot.
 	pages []Page[W]
 }
 
 // NewMemory constructs a snapshot of a single memory module, identified by its
 // module identifier and described by the given sequence of pages.
-func NewMemory[W any](moduleId uint, pages []Page[W]) Memory[W] {
+func NewMemory[W any](moduleId uint16, pages []Page[W]) Memory[W] {
 	return Memory[W]{moduleId, pages}
 }
 
 // ModuleId returns the module identifier of the memory captured by this
 // snapshot.
-func (p Memory[W]) ModuleId() uint {
+func (p Memory[W]) ModuleId() uint16 {
 	return p.moduleId
 }
